remove core-js and reorganize core and aria components
