+++ package.json
@@ -36,7 +36,6 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.16.5",
-    "core-js": "^3.20.0",
     "fastdom": "^1.0.10"
   },
   "browserslist": [
