+++ src/Aria/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "js-component-framework/Aria",
+  "private": true,
+  "main": "./index.js"
+}
