+++ src/core/Component.js
