+++ src/core/index.js
@@ -1,0 +1,7 @@
+import Component from './Component';
+import ComponentManager from './ComponentManager';
+
+export {
+  ComponentManager,
+  Component,
+};
